Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.6
